requirements.txt
asgiref==3.2.10
awscli==1.18.123
boto3==1.14.46
botocore==1.17.46
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0
django-allauth==0.42.0
django-environ==0.4.5
django-extensions==3.0.5
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-yasg==1.17.1
google-auth==1.20.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
phonenumbers==8.12.8
Pillow==7.2.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
supervisor==4.2.0
uritemplate==3.0.1
urllib3==1.25.10
